+++ machines/t500libre/httpd.nix
@@ -1,6 +1,5 @@
 # Configuration for virtual domains using apache2
 # @todo: move virtualhosts in a caddy environment
-# @todo: remove dslg from virtual hosts
 # @todo: update the archive project, can be moved to ideapad100 (becaues it already deals with python)
 # @todo: make an overall map of the services hosted at the gnu domain.
 {
@@ -109,21 +108,6 @@ in
             Header unset X-Frame-Options
           </LocationMatch>
         '';
-      };
-      # Debian Source List Generator for the masses.
-      "dslg.${gnu-domain}" = {
-        # forceSSL uses 302 Found redirects, using own 301 redirects in 'extraConfig'.
-        #addSSL = true;
-        enableACME = true;
-        #inherit acmeRoot;
-        forceSSL = true;
-        hostName = "dslg.${gnu-domain}";
-        serverAliases = [
-          "www.dslg.${gnu-domain}"
-        ];
-        documentRoot = "/var/www/dslg.${gnu-domain}/";
-        # serving php files as default.
-        locations."/".index = "index.php";
       };
       # Task planning tool.
       "do.${gnu-domain}" = {